+++ kicker.cpp
@@ -1202,7 +1202,7 @@ STDMETHODIMP Kicker::BallCntOver(int *pVal)
 
    if (g_pplayer)
    {
-      for (unsigned i = 0; i < g_pplayer->m_vball.size(); i++)
+      for (size_t i = 0; i < g_pplayer->m_vball.size(); i++)
       {
          Ball * const pball = g_pplayer->m_vball[i];
 
@@ -1232,13 +1232,13 @@ STDMETHODIMP Kicker::get_LastCapturedBall(IBall **pVal)
     }
 
     bool ballFound = false;
-    for (unsigned int i = 0; i < g_pplayer->m_vball.size(); i++)
+    for (size_t i = 0; i < g_pplayer->m_vball.size(); i++)
     {
         if (g_pplayer->m_vball[i] == m_phitkickercircle->m_lastCapturedBall)
-	{
+        {
             ballFound = true;
-	    break;
-	}
+            break;
+        }
     }
 
     if (!ballFound)
